+++ test/barcodes/test_barcodes.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from fpdf import FPDF
-from test.utilities import assert_pdf_equal
+from test.conftest import assert_pdf_equal
 
 HERE = Path(__file__).resolve().parent