+++ test/end_to_end_legacy/charmap/test_charmap.py
@@ -15,7 +15,7 @@ import pytest
 
 import fpdf
 from fpdf.ttfonts import TTFontFile
-from test.utilities import assert_pdf_equal
+from test.conftest import assert_pdf_equal
 
 HERE = Path(__file__).resolve().parent