+++ test/fonts/test_set_font.py
@@ -6,7 +6,7 @@ import pytest
 from fpdf import FPDF
 from fpdf.errors import FPDFException
 from fpdf.fonts import fpdf_charwidths
-from test.utilities import assert_pdf_equal
+from test.conftest import assert_pdf_equal
 
 HERE = Path(__file__).resolve().parent