+++ test/html/test_html.py
@@ -4,7 +4,7 @@ import pytest
 
 import fpdf
 from fpdf.html import px2mm
-from test.utilities import assert_pdf_equal
+from test.conftest import assert_pdf_equal
 
 
 HERE = Path(__file__).resolve().parent