+++ test/image/test_url_images.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 
 import fpdf
-from test.utilities import assert_pdf_equal
+from test.conftest import assert_pdf_equal
 
 HERE = Path(__file__).resolve().parent