+++ test/test_set_date.py
@@ -5,7 +5,7 @@ import pytest
 
 import fpdf
 from fpdf.errors import FPDFException
-from test.utilities import assert_pdf_equal
+from test.conftest import assert_pdf_equal
 
 
 HERE = Path(__file__).resolve().parent