+++ test/test_structure_tree.py
@@ -77,7 +77,9 @@ endobj"""
 
 
 def test_single_image_structure_tree():
-    images_alt_texts = (MarkedContent(1, 0, 0, "Image title", "Image description"),)
+    images_alt_texts = (
+        MarkedContent(1, 0, "/Figure", 0, "Image title", "Image description"),
+    )
     struct_builder = StructureTreeBuilder(images_alt_texts)
     assert (
         struct_builder.serialize(first_object_id=3)